+++ lib/stacks/garnet-ingestion/garnet-ingestion-stack.ts
@@ -69,7 +69,6 @@ export class GarnetIngestionStack extends NestedStack {
             architecture: Architecture.ARM_64,
             environment: {
               DNS_CONTEXT_BROKER: props.dns_context_broker,
-              URL_SMART_DATA_MODEL: Parameters.smart_data_model_url,
               AWSIOTTHINGTYPE: garnet_nomenclature.aws_iot_thing
             }
           }