+++ lib/stacks/garnet-iot/iot-group/index.ts
@@ -104,7 +104,6 @@ export class GarnetIotGroup extends Construct {
       architecture: Architecture.ARM_64,
       environment: {
         DNS_CONTEXT_BROKER: props.dns_context_broker,
-        URL_SMART_DATA_MODEL: Parameters.smart_data_model_url,
         AWSIOTTHINGTYPE: garnet_nomenclature.aws_iot_thing
       }
     })